requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
backtrader==1.9.76.123
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.1.7
djangorestframework==3.12.4
idna==2.10
idna-ssl==1.1.0
kiwisolver==1.3.1
matplotlib==3.2.2
multidict==5.1.0
mysql-connector-python==8.0.24
mysqlclient==1.4.6
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.2
protobuf==3.15.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
ta==0.7.0
telepot==12.7
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
